"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ module 2/Assignment-3/Solution.java
@@ -14,7 +14,7 @@ public class Solution {
 	/*
 	Need to write the power function and print the output.
 	*/
-	public static int power(int base,int exponent) {
+	public static long power(int base,int exponent) {
 		if (exponent!=0)
 			return (base*power(base, exponent-1));
 		else
